creation of a RESTApi singleton used to make calls to a REST API
