+++ src/main/scala/GAPY/GNS3_Manager.scala
@@ -6,7 +6,7 @@ import org.json.simple._
 class GNS3_Manager(val serverAddress:String) {
   
   def createProject(name : String) : ProjectManager = {
-    val returned = RESTCall("/v2/projects", "POST", "{\"name\": \"" + name + "\"}");
+    val returned = RESTApi.post("/v2/projects", "{\"name\": \"" + name + "\"}",serverAddress);
     val obj = JSONValue.parse(returned); 
     val jsonObj:JSONObject = obj.asInstanceOf[JSONObject];
     val projectId = jsonObj.get("project_id").asInstanceOf[String]
@@ -14,30 +14,12 @@ class GNS3_Manager(val serverAddress:String) {
   }
   
   def deleteProject(projectId: String) : GNS3_Manager = {
-    val returned = RESTCall("/v2/projects/" + projectId, "DELETE", "");
+    val returned = RESTApi.delete("/v2/projects/" + projectId, serverAddress);
     this
   }
   
   def getProjectId(name : String) : String = {
     return null
   }
-  
-  private def RESTCall(url:String, method:String, body:String) : String = {
-        var http = Http(serverAddress + url)
-        method match {
-            case "GET" => {
-              val response: HttpResponse[String] = http.asString
-              return response.body
-            }
-            case "POST" => {
-              val response: HttpResponse[String] = http.postData(body).asString    //.postForm(body).asString
-              return response.body
-            }
-            case "DELETE" => {
-              val response: HttpResponse[String] = http.method("delete").asString    //.postForm(body).asString
-              return response.body
-            }
-        }
-    } 
   
 }
